chore(cli): move package functions from lib to util
