+++ packages/cli/src/components/update-warning-table/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useRef } from "react";
 import Table from "ink-table";
 import { Box, Text, TextProps } from "ink";
-import { RefinePackageInstalledVersionData } from "src/lib/package-manager";
+import { RefinePackageInstalledVersionData } from "src/commands/check-updates";
 
 const columIndex = {
     name: 1,
